+++ engine/include/engine/initialize.h
@@ -23,6 +23,9 @@ std::pair<std::pair<std::shared_ptr<ral::cache::CacheMachine>,std::shared_ptr<ra
 void finalize();
 
 size_t getFreeMemory();
+void resetMaxMemoryUsed(int to = 0);
+size_t getMaxMemoryUsed();
+
 extern "C" {
 
 error_code_t initialize_C(uint16_t ralId,