+++ CHANGELOG.md
@@ -10,9 +10,9 @@
 
 ## Improvements
 - #1236 Moving code from header files to implementation files
+- #1257 Expose the reset max memory usage C++ API to python
 - #1256 Improve Logical project documentation
 - #1153 Enable warnings and fixes
-
 
 ## Bug Fixes
 - #1249 Fix compilation with cuda 11